Use int type np arrays

+++ convert_netlists_to_graph.py
@@ -128,7 +128,7 @@ node_count = adj_full.get_shape()[0]
 
 # Save features as a numpy matrix
 assert(len(feats) == node_count)
-feats_array = np.zeros( (node_count, len(feature_mapping)) )
+feats_array = np.zeros( (node_count, len(feature_mapping)),dtype=np.int32 )
 for node in feats:
     for feature in feats[node]:
         row = node
